Make file system accessing tests sequential
Unit tests which are accessing the file system are inherently non-thread safe. This can be an issue in the github pipeline where tests appear to be running in parallel. This could explain why tests fail on github but not locally.
We force those tests to run sequentially, expecting to prevent this failure.